+++ web/jest.config.js
@@ -3,12 +3,9 @@ module.exports = {
   rootDir: 'src',
   setupFiles: ['../polyfills.test.js', '../setup.enzyme.test.js'],
   setupFilesAfterEnv: ['../setup.rtl.test.js'],
-  moduleDirectories: [
-    'src',
-    'node_modules'
-  ],
+  moduleDirectories: ['src', 'node_modules'],
   moduleNameMapper: {
-    'apd-testing-library': '<rootDir>/shared/apd-testing-library.js'
+    '^apd-(.*)-library$': '<rootDir>/shared/apd-$1-library.js'
   },
   snapshotSerializers: ['enzyme-to-json/serializer'],
   transform: {
