+++ web/src/components/LoginForm.story.js
@@ -1,10 +1,14 @@
 import React from 'react';
 import { plain as LoginForm } from './LoginForm';
-import * as CardFormStories from './CardForm.stories';
+import * as CardFormStory from './CardForm.story';
 
 export default {
   title: 'LoginForm',
   component: LoginForm,
+  decorators: CardFormStory.decorators,
+  parameters: {
+    jest: ['LoginForm.test.js']
+  },
   argTypes: {
     cancelable: { control: 'boolean' },
     canSubmit: { control: 'boolean' },
@@ -20,59 +24,26 @@ export default {
 export const Basic = args => <LoginForm {...args} />;
 
 Basic.args = {
-  ...CardFormStories.Basic.args
+  ...CardFormStory.Basic.args
 };
-
-Basic.parameters = {
-  jest: ['LoginForm.test.js']
-};
-
-const Form = () => (
-  <div className="ds-u-margin-bottom--4">
-    <label
-      htmlFor="textfield"
-      id="textfield-label"
-      className="ds-c-label ds-u-margin-y--2 ds-u-font-weight--normal"
-    >
-      Enter some text in this field
-    </label>
-    <input
-      width="200px"
-      aria-labelledby="textfield-label"
-      className="ds-c-field ds-c-field--medium"
-      id="textfield"
-      type="text"
-      name="textfield"
-      defaultValue="Some Text"
-    />
-  </div>
-);
 
 export const TextField = args => (
   <LoginForm {...args}>
-    <Form />
+    <CardFormStory.Form />
   </LoginForm>
 );
 
 TextField.args = {
-  ...CardFormStories.Basic.args
-};
-
-TextField.parameters = {
-  ...Basic.parameters
+  ...Basic.args
 };
 
 export const TextFieldWithFooter = args => (
   <LoginForm {...args}>
-    <Form />
+    <CardFormStory.Form />
   </LoginForm>
 );
 
 TextFieldWithFooter.args = {
-  ...CardFormStories.Basic.args,
+  ...Basic.args,
   footer: <p className="ds-u-padding-top--2">This is a footer</p>
-};
-
-TextFieldWithFooter.parameters = {
-  ...Basic.parameters
 };
